pyproject.toml
[tool.poetry]
name = "lms-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
black = "^24.10.0"
isort = "^5.13.2"
django = "4.2.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
